fix: issue #649
Split summary into two functions summary and get_summary.
get_summary does the heavy lifting of creating the DataFrame.
summary now calls get_summary and displays the result using
dysh.util.core.show_DataFrame. This will show the summary
even if the call is not the last one in a cell.
It also adds an argument for the maximum number of rows
and hides the index by default.