Remove @classmethod + @property decorator combination in GBTGainCorrection.valid_scales
I assume this was made a property in order to decorate it, but this is causing issues on Python 3.14. So, I've converted it to a tuple and put the documentation into the class docstring. It's a tuple now so that we don't have to worry about its state mutating across instances