Expand use of fitsio partial file loads
Ensure that we are still caching the file in mem when accessed via astropy

Ensure that we are using that cache on subsequent row-level reads via fitsio. i.e. we don't want to hit disk to load 2 rows from the file if they are already available via the hdu in mem